backend/stock_monitoring/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dataclasses-json==0.6.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.2
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
llama-index-core==0.10.30
llama-parse==0.4.1
llamaindex-py-client==0.1.18
marshmallow==3.21.1
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.23.2
orjson==3.10.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pymongo==4.7.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
